Add 403 status for comment delete fucnctionality

+++ src/routers/comments-router.ts
@@ -20,7 +20,7 @@ commentsRouter.get("/:id", objectIdValidationMiddleware, async (req: Request, re
 })
 
 commentsRouter.put("/:id", authMiddleware , commentInputsValidation, inputValidationMiddleware, objectIdValidationMiddleware, async (req: Request, res: Response) => {
-    debugger
+
     const comment = await commentsService.findComment(new ObjectId(req.params.id))
     if (comment?.commentatorInfo.userId !== req.user?._id.toString()){
         res.sendStatus(403);
@@ -36,6 +36,12 @@ commentsRouter.put("/:id", authMiddleware , commentInputsValidation, inputValida
 
 
 commentsRouter.delete("/:id", authMiddleware, objectIdValidationMiddleware, async (req: Request, res: Response) => {
+    const comment = await commentsService.findComment(new ObjectId(req.params.id))
+    if (comment?.commentatorInfo.userId !== req.user?._id.toString()){
+        res.sendStatus(403);
+        return
+    }
+
     const isDeleted = await commentsService.deleteComment(new ObjectId(req.params.id));
     if (isDeleted) {
         res.sendStatus(HTTP_STATUSES.NO_CONTENT_204);
